Overhaul the pure-python StringIO wrapper.
There is now only one class, not three; several inefficiencies have been
corrected; and the broken-support-for-NaN workaround has been removed.
(This workaround was incomplete and appears to have always been somewhat
wrong, leading me to believe that it ceased to be relevant a long time
ago.  There are still test cases that should detect a problem.)