Fix several subtle bugs in BufferedByteStream and its tests.
One issue deserves special mention: truncate() is clearly documented in
both Python 2 and 3 as *not* changing the seek position.  In Python 2,
however, truncate(0) on a StringIO object *did* reset the seek position,
and the encoder and decoder tests ubiquitously assumed this behavior.
io.BytesIO in Python 3 behaves as documented, breaking a great many tests.