Fix set serialization.  The test suite is now green on Py3k.
In Python 3, there is no 'sets' module, so there was nothing to cause the
'_sets' adapter to be loaded, in turn causing test failures.  This would
also have affected Python 2 code that used only the built-in set types,
but it happens to be convenient to put the tests for built-in set types
and the tests for library set types in the same test script, so we were
getting away with it.

Also, sort sets on serialization and enforce this in the test suite.